all done except Prim
