+++ Graph/utilityTypes.h
@@ -1,5 +1,3 @@
-#pragma once
-
 #include <list>
 #include <vector>
 #include <unordered_map>
@@ -11,15 +9,16 @@ template<typename data_t, typename weight_t>
 struct Edge;
 template<typename data_t, typename weight_t>
 struct Vertex;
+
 template<typename data_t, typename weight_t>
 class Graph;
 
 /////////////define types////////////////////////
 #define umap std::unordered_map
 #define id_t std::string
-#define vertex_t Vertex<data_t, weight_t>   
-#define edge_t Edge<data_t, weight_t>       
-#define graph_t Graph<data_t, weight_t>     
+#define vertex_t Vertex<data_t, weight_t>
+#define edge_t Edge<data_t, weight_t>
+#define graph_t Graph<data_t, weight_t>
 #define dset_t DisjoinSetArray<id_t>
 #define Kpair_t std::pair<weight_t, edge_t*>
 #define Kqueue_t std::priority_queue<Kpair_t, std::vector<Kpair_t>, std::greater<Kpair_t>>
@@ -36,9 +35,9 @@ struct Edge {
         delete this;
     }
 
-	//~Edge(){
-		//killSelf();
-	//}
+    //~Edge(){
+    //killSelf();
+    //}
 
     Edge(vertex_t *v, weight_t w) {
         vertexes = v;
@@ -52,7 +51,7 @@ struct Vertex {
     data_t data;
     std::list<edge_t *> edges;
 
-	Vertex(){}
+    Vertex() {}
 
     void killSelf() {
         delete this;
@@ -73,28 +72,30 @@ struct Airport {
     string country;
     double latitude;
     double longitude;
-	Airport(){}
-	Airport(int i,string n,string ci, string co, double lat, double lon):
-		id(i),
-		name(n),
-		city(ci),
-		country(co),
-		latitude(lat),
-		longitude(lon)
-	{}
+
+    Airport() {}
+
+    Airport(int i, string n, string ci, string co, double lat, double lon) :
+            id(i),
+            name(n),
+            city(ci),
+            country(co),
+            latitude(lat),
+            longitude(lon) {}
 };
+
 ////////////////functions////////////////////////
 template<typename data_t, typename weight_t>
-void visit(umap<id_t,bool>& visited, umap<id_t, vertex_t*>& vertexes, id_t id){
-	visited[id]=true;
-	for(auto it = vertexes[id]->edges.begin(); it!=vertexes[id]->edges.end(); it++){
-		if(visited[(*it)->vertexes[1].id]==false)
-			visit(visited,vertexes,(*it)->vertexes[1].id);
-	}
+void visit(umap<id_t, bool> &visited, umap<id_t, vertex_t *> &vertexes, id_t id) {
+    visited[id] = true;
+    for (auto it = vertexes[id]->edges.begin(); it != vertexes[id]->edges.end(); it++) {
+        if (visited[(*it)->vertexes[1].id] == false)
+            visit(visited, vertexes, (*it)->vertexes[1].id);
+    }
 }
 
 template<typename data_t, typename weight_t>
-id_t getIdOf(vertex_t v){
-	return v.id;
+id_t getIdOf(vertex_t v) {
+    return v.id;
 }
